--- lib/project_types/script/layers/infrastructure/task_runner.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-module Script
-  module Layers
-    module Infrastructure
-      class TaskRunner
-        TASK_RUNNERS = {
-          "assemblyscript" => Infrastructure::AssemblyScriptTaskRunner,
-          "rust" => Infrastructure::RustTaskRunner,
-        }
-
-        def self.for(ctx, language, script_name)
-          raise Errors::TaskRunnerNotFoundError unless TASK_RUNNERS[language]
-          TASK_RUNNERS[language].new(ctx, script_name)
-        end
-      end
-    end
-  end
-end
