+++ test/project_types/script/layers/infrastructure/languages/task_runner_test.rb
@@ -2,16 +2,16 @@
 
 require "project_types/script/test_helper"
 
-describe Script::Layers::Infrastructure::TaskRunner do
+describe Script::Layers::Infrastructure::Languages::TaskRunner do
   describe "build" do
     let(:script_name) { "script_name" }
-    subject { Script::Layers::Infrastructure::TaskRunner.for(@context, language, script_name) }
+    subject { Script::Layers::Infrastructure::Languages::TaskRunner.for(@context, language, script_name) }
 
     describe "when the script language and compile type match an entry in the registry" do
       let(:language) { "assemblyscript" }
 
       it "should return the entry from the registry" do
-        Script::Layers::Infrastructure::AssemblyScriptTaskRunner
+        Script::Layers::Infrastructure::Languages::AssemblyScriptTaskRunner
           .expects(:new)
           .with(@context, script_name)
         subject
